ROll in Shady Island to use Workload over the CDK Stage
Workload is allegedly Stage, but independent of AWS's version of a CDK
Pipline, which is overly restrictive for our needs (one deployment per
account).

I'm doubtful.  The limitation of one per account seems to be imposed
purely by CDK Pipellines when used with Stage, but Stage is also
packaged in CDK Pipeline's namespace so...   Anyhow, isolating this
commit to capture the changes in isolation.